Dumper: divided into new classes Describer, Exposer & Renderer
