+++ src/tracy.php
@@ -13,7 +13,10 @@ require __DIR__ . '/Tracy/Bar/Bar.php';
 require __DIR__ . '/Tracy/Bar/DefaultBarPanel.php';
 require __DIR__ . '/Tracy/BlueScreen/BlueScreen.php';
 require __DIR__ . '/Tracy/Dumper/Value.php';
+require __DIR__ . '/Tracy/Dumper/Describer.php';
 require __DIR__ . '/Tracy/Dumper/Dumper.php';
+require __DIR__ . '/Tracy/Dumper/Exposer.php';
+require __DIR__ . '/Tracy/Dumper/Renderer.php';
 require __DIR__ . '/Tracy/Logger/ILogger.php';
 require __DIR__ . '/Tracy/Logger/FireLogger.php';
 require __DIR__ . '/Tracy/Logger/Logger.php';
